# apps/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entity.FoodItem:
    properties:
      barcode_upc:
        type: string
      brand_name:
        type: string
      calcium_mg_per_default_serving:
        type: number
      calories_per_default_serving:
        type: number
      carbs_grams_per_default_serving:
        type: number
      cholesterol_mg_per_default_serving:
        type: number
      created_at:
        type: string
      created_by_user_id:
        type: string
      fat_grams_per_default_serving:
        type: number
      fiber_grams_per_default_serving:
        type: number
      id:
        type: string
      iron_mg_per_default_serving:
        type: number
      is_verified:
        type: boolean
      name:
        type: string
      potassium_mg_per_default_serving:
        type: number
      protein_grams_per_default_serving:
        type: number
      saturated_fat_grams_per_default_serving:
        type: number
      serving_size_default_qty:
        type: number
      serving_size_default_unit:
        type: string
      sodium_mg_per_default_serving:
        type: number
      source:
        $ref: '#/definitions/entity.FoodItemSource'
      sugar_grams_per_default_serving:
        type: number
      trans_fat_grams_per_default_serving:
        type: number
      updated_at:
        type: string
      vitamin_a_mcg_per_default_serving:
        type: number
      vitamin_c_mg_per_default_serving:
        type: number
    type: object
  entity.FoodItemSource:
    enum:
    - user_created
    - system
    - api
    type: string
    x-enum-varnames:
    - FoodItemSourceUserCreated
    - FoodItemSourceSystem
    - FoodItemSourceAPI
  entity.MealFoodItem:
    properties:
      calories_consumed:
        type: number
      carbs_consumed:
        type: number
      fat_consumed:
        type: number
      food_item_id:
        type: string
      id:
        type: string
      logged_at:
        type: string
      meal_id:
        type: string
      protein_consumed:
        type: number
      quantity_consumed:
        type: number
      serving_unit_consumed:
        type: string
    type: object
...
  entity.UserMeal:
    properties:
      created_at:
        type: string
      custom_meal_name:
        type: string
      id:
        type: string
      meal_date:
        type: string
      meal_time:
        type: string
      meal_type:
        $ref: '#/definitions/entity.UserMealType'
      notes:
        type: string
      total_calories_consumed:
        type: number
      total_carbs_consumed:
        type: number
      total_fat_consumed:
        type: number
      total_protein_consumed:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  entity.UserMealType:
    enum:
    - breakfast
    - lunch
    - dinner
    - snack
    - other
    type: string
    x-enum-varnames:
    - UserMealTypeBreakfast
    - UserMealTypeLunch
    - UserMealTypeDinner
    - UserMealTypeSnack
    - UserMealTypeOther
...
  v1.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Fitness application API for tracking nutrition, meals, and products
  title: Rebound API
  version: "1.0"
paths:
...
  /food-items/{id}:
    delete:
      description: Delete a food item by its ID
      parameters:
      - description: Food item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete a food item
      tags:
      - food-items
    get:
      description: Get a food item by its ID
      parameters:
      - description: Food item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.FoodItem'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get a food item by ID
      tags:
      - food-items
    put:
      consumes:
      - application/json
      description: Update an existing food item
      parameters:
      - description: Food item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated food item details
        in: body
        name: foodItem
        required: true
        schema:
          $ref: '#/definitions/entity.FoodItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.FoodItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update a food item
      tags:
      - food-items
...
  /meals:
    post:
      consumes:
      - application/json
      description: Create a new meal for a user
      parameters:
      - description: Meal object
        in: body
        name: meal
        required: true
        schema:
          $ref: '#/definitions/entity.UserMeal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.UserMeal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Create a new meal
      tags:
      - meals
  /meals/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a meal by its ID
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete a meal
      tags:
      - meals
    get:
      consumes:
      - application/json
      description: Get a meal by its ID
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserMeal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get a meal by ID
      tags:
      - meals
    put:
      consumes:
      - application/json
      description: Update an existing meal
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      - description: Meal object
        in: body
        name: meal
        required: true
        schema:
          $ref: '#/definitions/entity.UserMeal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserMeal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update a meal
      tags:
      - meals
  /meals/{id}/food-items:
    get:
      consumes:
      - application/json
      description: Get all food items for a meal
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.MealFoodItem'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get meal food items
      tags:
      - meals
    post:
      consumes:
      - application/json
      description: Add a new food item to a meal
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      - description: Food item object
        in: body
        name: foodItem
        required: true
        schema:
          $ref: '#/definitions/entity.MealFoodItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.MealFoodItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Add food item to meal
      tags:
      - meals
  /meals/food-items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a food item from a meal
      parameters:
      - description: Food item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Delete meal food item
      tags:
      - meals
    put:
      consumes:
      - application/json
      description: Update an existing food item in a meal
      parameters:
      - description: Food item ID
        in: path
        name: id
        required: true
        type: string
      - description: Food item object
        in: body
        name: foodItem
        required: true
        schema:
          $ref: '#/definitions/entity.MealFoodItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.MealFoodItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Update meal food item
      tags:
      - meals
  /meals/user/{userID}:
    get:
      consumes:
      - application/json
      description: Get all meals for a specific user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.UserMeal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Get user meals
      tags:
      - meals
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"