# apps/backend/internal/controller/router/v1/meal.go
package v1

import (
	"github.com/gofiber/fiber/v2"
	"github.com/terrnit/rebound/backend/internal/usecase"
	"github.com/terrnit/rebound/backend/pkg/logger"
)

type MealRoutes struct {
	mealUC *usecase.MealUseCase
	log    logger.Interface
}

func NewMealRoutes(handler fiber.Router, uc *usecase.MealUseCase, l logger.Interface) {
	r := &MealRoutes{
		mealUC: uc,
		log:    l,
	}

	h := handler.Group("/meals")
	{
		h.Post("/", r.createMeal)
		h.Get("/:id", r.getMeal)
		h.Put("/:id", r.updateMeal)
		h.Delete("/:id", r.deleteMeal)
		h.Get("/user/:userID", r.getUserMeals)
		h.Post("/:id/food-items", r.addFoodItem)
		h.Get("/:id/food-items", r.getFoodItems)
		h.Put("/food-items/:id", r.updateFoodItem)
		h.Delete("/food-items/:id", r.deleteFoodItem)
	}
}

// @Summary Create a new meal
// @Description Create a new meal for a user
// @Tags meals
// @Accept json
// @Produce json
// @Param meal body entity.UserMeal true "Meal object"
// @Success 201 {object} entity.UserMeal
// @Failure 400 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals [post]
func (r *MealRoutes) createMeal(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}

// @Summary Get a meal by ID
// @Description Get a meal by its ID
// @Tags meals
// @Accept json
// @Produce json
// @Param id path string true "Meal ID"
// @Success 200 {object} entity.UserMeal
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals/{id} [get]
func (r *MealRoutes) getMeal(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}

// @Summary Update a meal
// @Description Update an existing meal
// @Tags meals
// @Accept json
// @Produce json
// @Param id path string true "Meal ID"
// @Param meal body entity.UserMeal true "Meal object"
// @Success 200 {object} entity.UserMeal
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals/{id} [put]
func (r *MealRoutes) updateMeal(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}

// @Summary Delete a meal
// @Description Delete a meal by its ID
// @Tags meals
// @Accept json
// @Produce json
// @Param id path string true "Meal ID"
// @Success 204 "No Content"
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals/{id} [delete]
func (r *MealRoutes) deleteMeal(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}

// @Summary Get user meals
// @Description Get all meals for a specific user
// @Tags meals
// @Accept json
// @Produce json
// @Param userID path string true "User ID"
// @Param page query int false "Page number" default(1)
// @Param pageSize query int false "Page size" default(10)
// @Success 200 {array} entity.UserMeal
// @Failure 400 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals/user/{userID} [get]
func (r *MealRoutes) getUserMeals(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}

// @Summary Add food item to meal
// @Description Add a new food item to a meal
// @Tags meals
// @Accept json
// @Produce json
// @Param id path string true "Meal ID"
// @Param foodItem body entity.MealFoodItem true "Food item object"
// @Success 201 {object} entity.MealFoodItem
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals/{id}/food-items [post]
func (r *MealRoutes) addFoodItem(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}

// @Summary Get meal food items
// @Description Get all food items for a meal
// @Tags meals
// @Accept json
// @Produce json
// @Param id path string true "Meal ID"
// @Success 200 {array} entity.MealFoodItem
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals/{id}/food-items [get]
func (r *MealRoutes) getFoodItems(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}

// @Summary Update meal food item
// @Description Update an existing food item in a meal
// @Tags meals
// @Accept json
// @Produce json
// @Param id path string true "Food item ID"
// @Param foodItem body entity.MealFoodItem true "Food item object"
// @Success 200 {object} entity.MealFoodItem
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals/food-items/{id} [put]
func (r *MealRoutes) updateFoodItem(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}

// @Summary Delete meal food item
// @Description Delete a food item from a meal
// @Tags meals
// @Accept json
// @Produce json
// @Param id path string true "Food item ID"
// @Success 204 "No Content"
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Router /meals/food-items/{id} [delete]
func (r *MealRoutes) deleteFoodItem(c *fiber.Ctx) error {
	// TODO: Implement
	return c.SendStatus(fiber.StatusNotImplemented)
}